Removed iframe from nav-global to fix warning in Chrome.
Chrome was preventing use of the theme - it looks like the src url used by
the iframe no longer exists (I get a 404) - and was displaying the message:

>> For security reasons, framing is not allowed

The message was displayed in an alert-style box. When I clicked 'Ok' to dismiss,
I got redirected to http://markdotto.github.com/github-buttons/github-btn.html?user=plusjade&repo=jekyll-bootstrap&type=fork&count=true
and got a 404.

Removing the iframe fixes this issue.